+++ src/address/address-model.ts
@@ -2,8 +2,9 @@
  * Copyright (c) TIKI Inc.
  * MIT license. See LICENSE file in root directory.
  */
+import { BlockPageModel } from '../block/block-page-model'
 
-type AddressModel = {
+export type AddressModel = {
   appId: string
   address: string
   blocks: BlockPageModel
