+++ src/index.ts
@@ -3,52 +3,38 @@
  * MIT license. See LICENSE file in root directory.
  */
 
-import AddressRepository from './address/address-repository'
-import AppRepository from './app/app-repository'
-import BlockRepository from './block/block-repository'
-import TxnRepository from './txn/txn-repository'
-import * as HttpHelpers from './utils/http-helpers'
+import L0Index from './l0-index'
+import { L0IndexConfig } from './l0-index-config'
+import { ErrorModel } from './utils/error-model'
+import { TxnGet } from './txn/txn-get'
+import { TxnModel } from './txn/txn-model'
+import { TxnContentsModel } from './txn/txn-contents-model'
+import { TxnContentsModelOwnership } from './txn/txn-contents-model-ownership'
+import { TxnContentSchemaEnum } from './txn/txn-content-schema-enum'
+import { TxnContentsModelConsent } from './txn/txn-contents-model-consent'
+import { BlockGet } from './block/block-get'
+import { BlockModel } from './block/block-model'
+import { BlockPageModel } from './block/block-page-model'
+import { AppModel } from './app/app-model'
+import { AddressModel } from './address/address-model'
+import { AddressGet } from './address/address-get'
+import { AddressPageModel } from './address/address-page-model'
 
-type L0IndexConfig = {
-  host: string
-}
-
-export default class L0Index {
-  appRepository: AppRepository
-  addressRepository: AddressRepository
-  blockRepository: BlockRepository
-  txnRepository: TxnRepository
-
-  constructor(config: L0IndexConfig) {
-    this.appRepository = new AppRepository(config.host)
-    this.addressRepository = new AddressRepository(config.host)
-    this.blockRepository = new BlockRepository(config.host)
-    this.txnRepository = new TxnRepository(config.host)
-  }
-
-  app = (
-    authorization: string,
-    appId: string,
-    page?: HttpHelpers.PageRequest
-  ): Promise<AppModel> => this.appRepository.get(authorization, appId, page)
-
-  address = (
-    authorization: string,
-    request: AddressGet,
-    page?: HttpHelpers.PageRequest
-  ): Promise<AddressModel> =>
-    this.addressRepository.get(authorization, request, page)
-
-  block = (
-    authorization: string,
-    request: BlockGet,
-    page?: HttpHelpers.PageRequest
-  ): Promise<BlockModel> =>
-    this.blockRepository.get(authorization, request, page)
-
-  transaction = (
-    authorization: string,
-    request: TxnGet,
-    page?: HttpHelpers.PageRequest
-  ): Promise<TxnModel> => this.txnRepository.get(authorization, request, page)
+export {
+  L0Index,
+  L0IndexConfig,
+  ErrorModel,
+  TxnModel,
+  TxnContentsModel,
+  TxnGet,
+  TxnContentsModelOwnership,
+  TxnContentSchemaEnum,
+  TxnContentsModelConsent,
+  BlockModel,
+  BlockPageModel,
+  BlockGet,
+  AddressModel,
+  AppModel,
+  AddressGet,
+  AddressPageModel,
 }
