+++ src/txn/txn-content-schema-enum.ts
@@ -3,7 +3,7 @@
  * MIT license. See LICENSE file in root directory.
  */
 
-enum TxnContentSchemaEnum {
+export enum TxnContentSchemaEnum {
   UNKNOWN = 'unknown',
   OWNERSHIP = 'ownership_nft',
   CONSENT = 'consent_nft',
