+++ src/txn/txn-model.ts
@@ -3,7 +3,12 @@
  * MIT license. See LICENSE file in root directory.
  */
 
-type TxnModel = {
+import { TxnContentsModel } from './txn-contents-model'
+import { TxnContentSchemaEnum } from './txn-content-schema-enum'
+import { TxnContentsModelOwnership } from './txn-contents-model-ownership'
+import { TxnContentsModelConsent } from './txn-contents-model-consent'
+
+export type TxnModel = {
   appId: string
   address: string
   block: string
